impl `ByteTokenizer`
It's a tokenizer that simply converts a string to a utf-8 byte array and
convert each byte to usize.

The biggest advantage of this tokenizer is that, unlike
`SimpleTokenizer`, you can reuse a model even though dataset changes.

I have tested it and it works well. I didn't commit the changes in
main.rs because I want to keep `SimpleTokenizer` as default.